refactor: Game -> GameController로 네이밍 변경
